# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Multilingual File Manager API",
        "version": "1.0.0",
        "description": "An API for managing user files with multilingual support, user authentication, CRUD operations for files, and a queuing system for file uploads."
    },
    "servers": [
        {
            "url": "http://localhost:5000",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/files": {
            "get": {
                "summary": "Get all files",
                "description": "Retrieves a list of all files uploaded by users.",
                "responses": {
                    "200": {
                        "description": "A list of files",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "files": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "id": { "type": "string" },
                                                    "name": { "type": "string" },
                                                    "path": { "type": "string" },
                                                    "mimeType": { "type": "string" },
                                                    "size": { "type": "integer" },
                                                    "description": { "type": "string" }
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": { "description": "Internal server error" }
                }
            }
        },
...
        "/api/i18n/language": {
            "get": {
                "summary": "Get current language",
                "description": "Returns the current language used by the user.",
                "responses": {
                    "200": {
                        "description": "Current language retrieved successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "language": { "type": "string", "example": "en" }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Set preferred language",
                "description": "Sets the preferred language for the user.",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "language": { "type": "string", "example": "fr" }
                                },
                                "required": ["language"]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Language updated successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": { "type": "string" },
                                        "language": { "type": "string" }
                                    }
                                }
                            }
                        }
                    },
                    "400": { "description": "Invalid language code." },
                    "500": { "description": "Internal server error." }
                }
            }
        }
    }
}